Update to Chromium revision 8cb76f56 (#307671) - Linux/Mac: Load V8 initial snapshot from external bin files (see http://crbug.com/421063). - Windows: 8.1 SDK is now required (issue #1470). - Remove CefDOMDocument::GetSelectionStartNode and GetSelectionEndNode methods(see https://codereview.chromium.org/763043004). - Rename internal namespace used in base headers to avoid linker conflicts with cef_sandbox.lib.
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@1956 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ libcef/browser/chrome_browser_process_stub.cc
@@ -122,7 +122,7 @@ GpuModeManager* ChromeBrowserProcessStub::gpu_mode_manager() {
 void ChromeBrowserProcessStub::CreateDevToolsHttpProtocolHandler(
     chrome::HostDesktopType host_desktop_type,
     const std::string& ip,
-    int port) {
+    uint16 port) {
 }
 
 unsigned int ChromeBrowserProcessStub::AddRefModule() {
