+++ libcef/browser/chrome_browser_process_stub.h
@@ -53,7 +53,7 @@ class ChromeBrowserProcessStub : public BrowserProcess {
   void CreateDevToolsHttpProtocolHandler(
       chrome::HostDesktopType host_desktop_type,
       const std::string& ip,
-      int port) override;
+      uint16 port) override;
   unsigned int AddRefModule() override;
   unsigned int ReleaseModule() override;
   bool IsShuttingDown() override;
